+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     * ip -> url conversion generates a proper PTR hostname
     * wrong_url protocol case insensitive (accepts "hXXp")
     * Unix time is reasonably recognized as a timestamp
+    * cc_contact converts abusemail to value given by contacts_cc.csv
 * dialog
     * edit default e-mail templates from config submenu
     * delimiter now may contain tab character, input by "\t" or "tab"